requirements.txt
matplotlib==3.8.0
Pillow==10.0.1
torch==2.2.1
torchvision==0.17.1
tqdm==4.66.2
